feat(deb/extract): preserve parent dir attributes
The purpose of parent directory handling in deb/extract.go is to create
parent directories of requested paths with the same attributes (only
mode as of now) that appear in the package tarball. However, the current
implementation is not correct when glob paths are requested.

In what follows, parent directory refers to a directory path that is not
explicitly requested for extraction, but that is the parent of other
paths that are requested for extraction, and so it is assumed to be
implicitly requested for extraction.

Currently, whether a package path should be extracted is determined by
the shouldExtract() function that iterates over requested paths and for
each checks whether it matches the package path if it's glob, or if it's
non-glob, whether it equals the package path or whether some of its
target paths have the package path as the parent.

There is a lacking with this implementation:

   It only checks whether a package path is the parent of any target
   path of a requested non-glob path. It does not, and probably even
   cannot, check whether it is the parent of a requested glob path.

This commit addresses this issue. Track package files' info and a list
of extracted paths as we iterate reading the contents of the package.
Afterwards, iterate over the extractd paths in order and change the mode
of parent directories, if those directories appear in the package.

Co-authored-by: Tomáš Virtus <tomas.virtus@canonical.com>