feat: support for multiple signatures in InRelease
The archive InRelease files may have multiple signatures. This commit
adds the support to verify the signatures.

Currently, similar to apt [1], chisel will verify at least one signature
and move on. This is quite unlike ``gpg --verify`` which ensures the
verification of all signatures.

This commit also updates the real-archive tests for a wider range of
support.

[1]  https://salsa.debian.org/apt-team/apt/-/blob/4e344a4c1d2862b7cdb900a20222bc22ac5edcf7/methods/gpgv.cc#L553-557