fix: address review related to tests and nitpicks
This commit removes the Ubuntu Archive Automatic Signing Key 2012 from
the live archive tests. It also refactors a bunch of codes here and
there. Most notably, this commit introduces the
``setup.DecodeAnySignature`` function which verifies if any signature in
a document is valid using a given set of public keys.