refactor: make package archive retrieval modular
In slicer.go, we figure out which archive will be used for a package to
get its contents. This has been so far written inside ``slicer.Run()``.
However, this simple logic may be enhacned later with the introduction
of multiple archives, priorities and so on. Additionally, this may be
used in other places in the codebase too. Hence, this commit moves the
logic to a new method named ``slicer.PackageArchive()`` and refactors
the calls in ``slicer.Run()``.