fix: "generate" paths cannot conflict with any path
A path specified with the "generate" keyword is supposed to generate
stuff. In itself, these are artificial paths. Which means you can have
the same path with same properties specified across packages. This
commit supports that as a first.

Additionally, since the "generate" paths are essentially globs, it
should not conflict with any other path. This commit ensures that too.