feat: add "generate" keyword to paths
Certain paths can use the new "generate" keyword to generate files
including the chisel DB itself. There are a few restrictions on the
paths' format however, when using the "generate" keyword. Paths must
have the following format:

    /absolute/path/to/dir/**

The path must be absolute, must have trailing /** and must not have any
other wildcard character other than the trailing **. The indication is
that any files generated will match the above path pattern.

Currently, the "generate" keyword only accepts one value: "manifest",
other than the empty default value "". The "generate: manifest" value
instructs chisel to generate the Chisel DB (or, manifest) at the
specified location.

Multiple paths can utilise this keyword to generate the Chisel DB at
multiple locations. Ideally, per discussions, the "base-files_chisel"
slice should have a "/var/lib/chisel/**" path which would do so. But
chisel does not put any restrictions on the location or the slice.

The Chisel DB will only be generated at the specified location(s) if the
path owning slices are installed. For example, if the
"base-files_chisel" slice specifies a path with "generate: manifest",
but the "base-files_chisel" slice is not being installed by the cut
command, the Chisel DB will not be generated at that path for that
slice.

These "generate" paths are similar to globs, and will conflict with
paths that match the pattern.

Note that this commit does not add the support for creation of the
Chisel DB. The scope of this commit is limited to the support for
specifying the "generate" keyword.