feat: support multiple archives with "priority" field
This commit adds support for fetching packages from multiple archives.
It introduces a new field "archives.<archive-name>.priority" which takes
in an integer and specifies the priority of a certain archive. This
value is considered when there are multiple archives. A package is
fetched from the archive with highest priority, unless the slice
definition file of that package specifies a particular archive using the
"archive" field.

TODO:
  a. Unit tests for cmd_cut.go, focusing on the archive selection for
     each package ("selectPkgArchive"). The PR for Chisel DB is moving
     code around and it seems best at the moment to wait for that change
     to be somewhat finalized.

Reference: Specification RK018.