feat: support "pro" keyword in archives
In chisel.yaml, archive definitions can now use the "pro" value to
specify Ubuntu Pro archives. The ``archives.<archive>.pro`` value
currently accepts the following values: "fips", "fips-updates", "apps"
and "infra". Any other values are ignored (with a log message) and reset
to "" (empty).

The following table shows the base repository URL for each of the
supported pro values:

    fips            https://esm.ubuntu.com/fips/ubuntu
    fips-updates    https://esm.ubuntu.com/fips-updates/ubuntu
    apps            https://esm.ubuntu.com/apps/ubuntu
    infra           https://esm.ubuntu.com/infra/ubuntu

The actual mechanism for fetching packages from these archives are not
in the scope of this commit and will follow in a later commit.

While using multiple archives, the "priority" values of these Ubuntu Pro
archives should be consistent with the following table:

    fips            20
    fips-updates    21
    apps            16
    infra           15

Note that chisel does not cross-check nor ensure the consistency of
"priority" with "pro" values.

Reference: Specification RK018.