feat: support multiple archives with "priority"
This commit adds support for fetching packages from multiple archives.
It introduces a new field ``archives.<archive-name>.priority`` which
takes in a signed integer and specifies the priority of a certain
archive. This value is particularly useful when there are multiple
archives. A package is fetched from the archive with highest priority
regardless of the "default" archive, unless the slice definition file of
that package specifies a particular archive using the "archive" field.

Example chisel.yaml:

    format: v1
    archives:
	foo:
	    version: 22.04
	    components: [main, universe]
	    priority: 20
	    public-keys: [..]
	bar:
	    version: 22.04
	    components: [main]
	    default: true
	    priority: 10
	    public-keys: [..]

In the above example, if a package exists in both of the archives, the
package will be fetched from archive "foo" since it has higher priority.
Note that, archive "bar" is the default. However, if in a particular
package such as below, the "archive" field is used to specify a
particular archive, that particular archive will be fetched from.

    package: test-package
    archive: bar
    slices:
    	...

The above package will be fetched from archive "bar", regardless of
priority.

Reference: Specification RK018.