test: real archive tests for Ubuntu Pro archives
To run the real archive tests for Ubuntu Pro archives, the host machine
needs to be Pro-enabled first.

    sudo pro attach <pro-token> --no-auto-enable

There are tests for fips, fips-updates, esm-apps, esm-infra archives.
Thus, the respective services have to be enabled as well to generate the
APT credentials.

    sudo pro enable fips fips-updates esm-apps esm-infra --assume-yes

We need to pass the --pro-archive flag in addition to the --real-archive
flag to run the tests for Pro archives.

    go test ./internal/archive/ --real-archive --pro-archive

Note that the credentials file(s) in /etc/apt/auth.conf.d (or
CHISEL_AUTH_DIR) may require root permission.

test: pass missing pro field in archive.Options